requirements.txt
torch>=2.1,<3
torchvision>=0.16,<1
streamlit>=1.29,<2
pillow>=10,<11
numpy>=1.24,<2
scikit-learn>=1.3,<2
